+++ app/src/main/AndroidManifest.xml
@@ -18,9 +18,6 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31">
         <activity
-            android:name=".player.ui.PlayerActivity"
-            android:exported="false" />
-        <activity
             android:name=".main.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustNothing">
