+++ app/src/main/res/navigation/navigation_graph.xml
@@ -9,11 +9,7 @@
         android:id="@+id/favoritesFragment"
         android:name="com.practicum.playlistmaker.favorites.ui.FavoritesFragment"
         android:label="fragment_favorites"
-        tools:layout="@layout/fragment_favorites" >
-        <action
-            android:id="@+id/action_favoritesFragment_to_playerActivity"
-            app:destination="@id/playerActivity" />
-    </fragment>
+        tools:layout="@layout/fragment_favorites" />
     <fragment
         android:id="@+id/playlistsFragment"
         android:name="com.practicum.playlistmaker.media.ui.PlaylistsFragment"
@@ -25,8 +21,8 @@
         android:label="fragment_search"
         tools:layout="@layout/fragment_search">
         <action
-            android:id="@+id/action_searchFragment_to_playerActivity"
-            app:destination="@id/playerActivity" />
+            android:id="@+id/action_searchFragment_to_playerFragment"
+            app:destination="@id/playerFragment" />
     </fragment>
     <fragment
         android:id="@+id/mediaFragment"
@@ -36,20 +32,27 @@
         <action
             android:id="@+id/action_mediaFragment_to_playlistCreationFragment"
             app:destination="@id/playlistCreationFragment" />
+        <action
+            android:id="@+id/action_mediaFragment_to_playerFragment"
+            app:destination="@id/playerFragment" />
     </fragment>
     <fragment
         android:id="@+id/settingsFragment"
         android:name="com.practicum.playlistmaker.settings.ui.SettingsFragment"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
-    <activity
-        android:id="@+id/playerActivity"
-        android:name="com.practicum.playlistmaker.player.ui.PlayerActivity"
-        android:label="PlayerActivity"
-        tools:layout="@layout/activity_player" />
     <fragment
         android:id="@+id/playlistCreationFragment"
         android:name="com.practicum.playlistmaker.playlists.ui.PlaylistCreationFragment"
         android:label="fragment_playlist_creation"
         tools:layout="@layout/fragment_playlist_creation" />
+    <fragment
+        android:id="@+id/playerFragment"
+        android:name="com.practicum.playlistmaker.player.ui.PlayerFragment"
+        android:label="fragment_player"
+        tools:layout="@layout/fragment_player" >
+        <action
+            android:id="@+id/action_playerFragment_to_playlistCreationFragment"
+            app:destination="@id/playlistCreationFragment" />
+    </fragment>
 </navigation>
